requirements.txt
#python 3.11.2
fastapi==0.103.1
matplotlib==3.8.0
numpy==1.26.0
psycopg2==2.9.7
pydub==0.25.1
python-multipart==0.0.6
scipy==1.11.2
uvicorn==0.23.2
python-decouple